% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_helpers.R
\name{default_params}
\alias{default_params}
\title{Load default parameters for Gourmex}
\usage{
default_params(grouped = FALSE)
}
\arguments{
\item{grouped}{If TRUE, returns additional parameters for comparative analysis.}
}
\value{
A full set of parameters to make running the Gourmex process.
}
\description{
Load default parameters for Gourmex
}
\examples{
params = default_params()
gourmex = mount(c('Sepal.Length', 'Petal.Width'), 'Species', params = params)
}
