% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.R
\name{dispatch}
\alias{dispatch}
\title{Dispatch}
\usage{
dispatch(gourmex)
}
\arguments{
\item{gourmex}{Gourmex ready for dispatch (shaping done).}
}
\value{
Gourmex ready to serve
}
\description{
Dispatch
}
\examples{
data = tibble::rownames_to_column(iris, 'fid')
gourmex = mount(c('Sepal.Length', 'Species'), id_cols='fid', params=default_params())
gourmex = ingest(gourmex, data)
gourmex = digest(gourmex)
gourmex = shape(gourmex)
gourmex = dispatch(gourmex)

}
