% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_helpers.R
\name{format_categorical}
\alias{format_categorical}
\title{Translate terms following a given dictionary.}
\usage{
format_categorical(
  V,
  df_dict,
  .fn = NULL,
  from = "id",
  to = "label",
  default = NULL,
  unmatched = NULL
)
}
\arguments{
\item{V}{A character vector containing terms to translate.}

\item{df_dict}{A data frame used as a dictionary (columns=var, rname, label, group).}

\item{.fn}{is applied finally on the translated terms.}

\item{from}{Reference for origin column.}

\item{to}{Reference for destination column.}

\item{default}{Default value if no correspondance was found in \emph{df_dict}. If NULL,
keep the previous values.}
}
\value{
A character vector with the translated terms.
}
\description{
Translate terms following a given dictionary.
}
\examples{
df_dict = tibble::tibble(low=letters, up=LETTERS)
V = sample(letters, 5, replace=TRUE)
format_categorical(V, df_dict, from='low', to='up')
}
