% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_helpers.R
\name{format_list}
\alias{format_list}
\title{Apply a format function on a list-column, as if it was a unnested.}
\usage{
format_list(L, format_func, .fn = NULL)
}
\arguments{
\item{L}{The list whose values must be formatted.}

\item{format_func}{The format function to apply on the unnested data.}

\item{.fn}{A post-process function to apply on row elements. Ignored if NULL.}
}
\value{
A formatted list
}
\description{
Apply a format function on a list-column, as if it was a unnested.
}
\examples{
L = list(A=c(12,3.1), B=NULL, C=c(1202.435), D=c(NA, 9, 11.7))
format_list(L, \(x) format_numbers(x, big.mark=',', lower_bound=10))
}
