% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_helpers.R
\name{format_perc}
\alias{format_perc}
\title{Tool to display percentages}
\usage{
format_perc(x, precision = 1, digits = 0, sum_to_100 = F, add_sign = F)
}
\arguments{
\item{x}{A numeric vector with ratios as values.}

\item{precision}{Passed to \code{round_numbers}. Applied on values multiplied by 100.}

\item{digits}{Passed to \code{format_numbers}.}

\item{sum_to_100}{If TRUE, apply the best rounding adjustments to assure a sum equal to 100\%.}

\item{add_sign}{If TRUE, basically paste a "\%" sign.}
}
\value{
A character vector with formatted percentages
}
\description{
Tool to display percentages
}
\examples{
x = c(0.12,0.346,0.534)
format_perc(x, digits=1, add_sign=TRUE)
}
