% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{ingest}
\alias{ingest}
\title{Ingest data: Prepare for analysis}
\usage{
ingest(gourmex, data)
}
\arguments{
\item{gourmex}{Mounted Gourmex object ready for analysis.}

\item{data}{The data to ingest.}
}
\value{
Gourmex object ready for digestion.
}
\description{
During ingestion the algorithm will process each variable to analyse by
starting evaluating its main type (categorical, logical, numerical), then
preparing homogeneous dataframes for stats.

The \code{logical} type usually refers to the same type as \code{categorical}, however
can be displayed in a different way (single line).
}
\examples{
data = tibble::rownames_to_column(iris, 'fid')
gourmex = mount(c('Sepal.Length', 'Species'), id_cols='fid')
gourmex = ingest(gourmex, data)
}
