% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{stat_n_avail}
\alias{stat_n_avail}
\title{Extend the statistics with the number of available data as a column named \code{stat_name}}
\usage{
stat_n_avail(gourmex, var, stat_name = "n_avail", by = NULL, keep_na_rows = T)
}
\arguments{
\item{var}{The column name of \code{df} where nan values can be detected.}

\item{stat_name}{indicates how the column reporting the number of available data should be named.}

\item{by}{is used as grouping variables}

\item{df}{A \code{report_df} object.}
}
\value{
An extended \code{report_df} object.
}
\description{
Extend the statistics with the number of available data as a column named \code{stat_name}
}
