% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{stat_pvalue}
\alias{stat_pvalue}
\title{Extend the statistics with the significance}
\usage{
stat_pvalue(gourmex, var, stat_name = "p_value", by = NULL, strata_var = NULL)
}
\arguments{
\item{gourmex}{A \code{gourmex} object.}

\item{var}{The variable used to evaluate p-value.}

\item{stat_name}{Indicates how the p-values columns should be named.}

\item{by}{Is used as grouping variables.}

\item{strata_var}{is used to separate multiple comparisons.}
}
\value{
A dataframe with the required p-values.
}
\description{
Extend the statistics with the significance
}
