% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{stat_total}
\alias{stat_total}
\title{Extend the statistics with a total column named \code{total_col}.}
\usage{
stat_total(gourmex, var, stat_name = "N", by = NULL)
}
\arguments{
\item{stat_name}{indicates how the new column should be named.}

\item{by}{is used as grouping variables.}

\item{df}{A \code{report_df} data frame.}
}
\value{
An extended \code{report_df} object.
}
\description{
Extend the statistics with a total column named \code{total_col}.
}
