% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yummystats-package.R
\docType{package}
\name{yummystats-package}
\alias{yummystats}
\alias{yummystats-package}
\title{yummystats: Companion to Render Tasty Aggregated Stats From Your Delicious Raw Data.}
\description{
This package provides a toolbox to quickly summarize your data in a nice custom layout. It starts by applying statistical functions which depend on the variable type (categorical, logical, numerical), before formatting values and displaying them as you wish.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Thorry13/yummystats}
  \item Report bugs at \url{https://github.com/Thorry13/yummystats/issues}
}

}
\author{
\strong{Maintainer}: Remy Dumas \email{remy.dumas84@gmail.com}

Authors:
\itemize{
  \item Remy Dumas \email{remy.dumas84@gmail.com}
}

}
\keyword{internal}
