% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digest.R
\name{digest}
\alias{digest}
\title{Run Gourmex digestion}
\usage{
digest(gourmex)
}
\arguments{
\item{gourmex}{A Gourmex object ready for digestion (ingestion done).}
}
\value{
A Gourmex ready for shaping.
}
\description{
At this state, the Gourmex object is sufficiently set up and can process the
statistical analysis.

It saves the results internally so it can be used later with minimal arguments.
}
\examples{
data = tibble::rownames_to_column(iris, 'fid')
gourmex = mount(c('Sepal.Length', 'Species'), id_cols='fid', params=default_params())
gourmex = ingest(gourmex, data)
gourmex = digest(gourmex)

}
