% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_helpers.R
\name{format_numbers}
\alias{format_numbers}
\title{Basic tool to format numbers}
\usage{
format_numbers(x, big.mark = "", digits = 0, lower_bound = NULL)
}
\arguments{
\item{x}{A numeric vector to format.}

\item{big.mark}{Passed to \code{formatC}.}

\item{digits}{Passed to \code{formatC}.}

\item{lower_bound}{If not NULL, used as a threshold to hide values below.}
}
\value{
A character vector with formatted numbers
}
\description{
Basic tool to format numbers
}
\examples{
x = c(13,52.1,2300,4)
format_numbers(x, big.mark=',', digits=2, lower_bound=10)
}
