% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{ingest}
\alias{ingest}
\title{Initialize}
\usage{
ingest(data, vars, id, group, params)
}
\arguments{
\item{data}{Data to digest.}

\item{vars}{Columns to analyse.}

\item{id}{A set of columns that uniquely identify each observation.}

\item{group}{A set of columns used to group observations}

\item{params}{A set of parameters for the analysis}
}
\description{
Initialize
}
