% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_helpers.R
\name{round_numbers}
\alias{round_numbers}
\title{This is an advanced function to round numbers.
It can apply a "smart rounding" taking all values into account instead of
considering them independently to make it consistent with the total (e.g. percentages).
There are optional parameters to keep raw values for low numbers, which might be useful
in this environment of functions.}
\usage{
round_numbers(
  V,
  precision = 1,
  total_consistency = FALSE,
  keep_low_numbers = FALSE,
  low_limit = NULL
)
}
\arguments{
\item{V}{A numeric vector whose values need to be rounded}

\item{precision}{If NULL, no rounding is applied.}

\item{total_consistency}{If TRUE, find the best adjustments to make the sum of rounded values
equal to the rounded sum.}

\item{keep_low_numbers}{If TRUE, avoid rounding low numbers, according to \code{low_limit} argument.}

\item{low_limit}{Values below this threshold are considered as low. If NULL, \code{low_limit} is set as
\code{precision}.}
}
\value{
A numeric vector with rounded values
}
\description{
This is an advanced function to round numbers.
It can apply a "smart rounding" taking all values into account instead of
considering them independently to make it consistent with the total (e.g. percentages).
There are optional parameters to keep raw values for low numbers, which might be useful
in this environment of functions.
}
\examples{
V = c(3.3,33.3,13.4)
round_numbers(V)
round_numbers(V, precision=10)
round_numbers(V, total_consistency=T)
round_numbers(V, precision=5, keep_low_numbers=T)
round_numbers(V, precision=5, keep_low_numbers=T, low_limit=3)
}
