% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{shape}
\alias{shape}
\title{Gourmex shaping.}
\usage{
shape(gourmex)
}
\value{
A Gourmex with shaped statistical information (ready for dispatch).
}
\description{
Once the statistics are computed and stored inside a Gourmex,
they usually need to be formatted to be more human-friendly.

This can be done through this shaping operation. It will transform each
statistical column independently, usually into character vectors.
}
