% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{stat_n_miss}
\alias{stat_n_miss}
\title{Extend the statistics with the number of missing data as a column named \code{stat_name}}
\usage{
stat_n_miss(gourmex, var, stat_name = "n_miss", by = NULL)
}
\arguments{
\item{gourmex}{A \code{gourmex} object.}

\item{var}{The column name of \code{df} where nan values can be detected.}

\item{stat_name}{indicates how the column reporting the number of available data should be named.}

\item{by}{Is used as grouping variables}
}
\value{
A dataframe with the required missing values.
}
\description{
Extend the statistics with the number of missing data as a column named \code{stat_name}
}
