% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{stat_n_miss}
\alias{stat_n_miss}
\title{Extend the statistics with the number of missing data as a column named \code{stat_name}}
\usage{
stat_n_miss(gourmex, var, stat_name = "n_miss", by = NULL)
}
\arguments{
\item{var}{The column name of \code{df} where nan values can be detected.}

\item{by}{is used as grouping variables}

\item{df}{A \code{report_df} object.}

\item{count_col}{The name of the column to count non-nan values on.}

\item{missing_col}{indicates how the column reporting the number of missing data should be named.}
}
\description{
Extend the statistics with the number of missing data as a column named \code{stat_name}
}
