% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{stat_numeric}
\alias{stat_numeric}
\title{Extend statistics for numerical variables, with the following values :
mean, std, min, Q1, median, Q3, max}
\usage{
stat_numeric(gourmex, var, stat_name, suffix = NULL, by = NULL)
}
\arguments{
\item{gourmex}{A \code{gourmex} object.}

\item{var}{A numerical variable name.}

\item{stat_name}{Name used to store the results.}

\item{suffix}{Used to append a suffix on numerical statistics names.}

\item{by}{Used to group results.}
}
\value{
A dataframe with the numerical stats.
}
\description{
Extend statistics for numerical variables, with the following values :
mean, std, min, Q1, median, Q3, max
}
