% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{stat_numeric}
\alias{stat_numeric}
\title{Extend statistics for numerical variables, with the following values :
mean, std, min, Q1, median, Q3, max}
\usage{
stat_numeric(gourmex, var, stat_name, suffix = NULL, by = NULL)
}
\arguments{
\item{suffix}{append suffix to the}

\item{df}{A \code{report_df} object}
}
\description{
Extend statistics for numerical variables, with the following values :
mean, std, min, Q1, median, Q3, max
}
