% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{stat_perc}
\alias{stat_perc}
\title{Extend the statistics with percentage column named \code{stat_name}.}
\usage{
stat_perc(
  gourmex,
  var,
  stat_name = "p",
  numerator_col = "n",
  denominator_col = NULL,
  by = NULL
)
}
\arguments{
\item{gourmex}{A \code{gourmex} object.}

\item{stat_name}{indicates how the percentage column should be named.}

\item{numerator_col}{The numerator.}

\item{denominator_col}{The denominator. If NULL, the sum of the \code{count_col} column is used as \code{total_col}.}

\item{by}{is used as grouping variables.}
}
\value{
A dataframe with the required percentages.
}
\description{
Extend the statistics with percentage column named \code{stat_name}.
}
