% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{stat_pvalue}
\alias{stat_pvalue}
\title{Extend the statistics with the significance}
\usage{
stat_pvalue(gourmex, var, stat_name = "p_value", by = NULL, strata_var = NULL)
}
\arguments{
\item{stat_name}{indicates how the p-values columns should be named.}

\item{by}{is used as grouping variables.}

\item{df}{A \code{report_df} object}
}
\value{
An extended \code{report_df} object.
}
\description{
Extend the statistics with the significance
}
