% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{stat_total}
\alias{stat_total}
\title{Calculate counts of a level for a given variable.}
\usage{
stat_total(gourmex, var, stat_name = "N", by = NULL)
}
\arguments{
\item{gourmex}{A \code{gourmex} object.}

\item{var}{The variable to compute statistic on.}

\item{stat_name}{Indicates how the new column should be named.}

\item{by}{Is used as grouping variables.}
}
\value{
A dataframe with the required counts.
}
\description{
Calculate counts of a level for a given variable.
}
